Added svelte-add-vitest
Noticed that there was no standard tool for adding Vitest to Sveltekit projects. Which is weird because it seems like the way to go with Svelte. So, I created a quick utility based on svelte-add-jest to add vitest instead, and I think that will be helpful for people new to Svelte.